[proj-47] refactor: ConsultingTabLayout 파일명 변경
